Update assemble-routed-files.sh
Permuted the label of $1 and $2 in "usage" to make them consistent with the CONFIG and ROUTED_FILES variables in the rest of the script. Unproper referencing of these variables turns into aberrant filenames/etc....with this script not doing the job at the end.
It would be great to update the Readme accordingly (same typo).